requirements.txt
absl-py==2.1.0
alembic==1.14.0
APScheduler==3.11.0
asgiref==3.8.1
attrs==24.2.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==5.1.3
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.2
flatbuffers==24.3.25
fonttools==4.54.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.4
kiwisolver==1.4.7
lxml==5.3.0
Mako==1.3.6
MarkupSafe==3.0.2
matplotlib==3.9.2
mediapipe==0.10.14
ml_dtypes==0.5.0
MouseInfo==0.1.3
numpy==2.1.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
protobuf==4.25.5
PyAutoGUI==0.9.54
pycparser==2.22
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
qrcode==8.0
requests==2.32.3
scipy==1.14.1
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
speedtest-cli==2.1.3
SQLAlchemy==2.0.36
sqlparse==0.5.2
tenacity==9.0.0
termcolor==2.5.0
tk==0.1.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
